editing unauthorized_users_may_not_delete_threads
working on Deleted Thread by authorized users

+++ app/Http/Controllers/ThreadsController.php
@@ -60,11 +60,22 @@ class ThreadsController extends Controller
      */
     public function destroy($channel, Thread $thread)
     {
+        if ( $thread->user_id != auth()->id() )
+        {
+            if ( request()->wantsJson() )
+            {
+                return response([ 'status' => 'Permission Denied', 403 ]);
+            }
+            return redirect('login');
+        }
+
         $thread->delete();
         if ( request()->wantsJson() ) {
             return response([], 204);
         }
         return redirect('/threads');
+
+
     }
 
     public function store()
